+++ flooding_worker/urls.py
@@ -8,20 +8,32 @@ from django.contrib import admin
 from django.http import HttpResponseServerError
 from django.template import Context
 from django.template import loader
+from flooding_worker.views import (
+    WorkflowTasksView, WorkflowsView, LoggingView)
 
 admin.autodiscover()
 handler404  # pyflakes
 
 urlpatterns = patterns(
     '',
-    url(r'^log/$', 'flooding_worker.views.homepage',
-        name='flooding_worker.homepage'),
-    url(r'^log/customer/(?P<customer_id>\d+)/$',
-     'flooding_worker.views.homepage',
-        name='flooding_worker-workflows'),
-    url(r'^log/customer/(?P<customer_id>\d+)/workflow/(?P<workflow_id>\d+)/$',
-     'flooding_worker.views.homepage',
-        name='flooding_worker-loggings'),
+    url(r'^scenario/(?P<scenario_id>\d*)/$',
+        WorkflowsView.as_view(),
+        name='flooding_worker_scenario'),
+    url(r'^workflow/(?P<workflow_id>\d+)/tasks/$',
+        WorkflowTasksView.as_view(),
+        name='flooding_worker_workflow_task'),
+    url(r'^scenario/(?P<scenario_id>\d*)/workflow/(?P<workflow_id>\d+)/loggings/$',
+        LoggingView.as_view(),
+        name='flooding_worker_workflow_logging'),
+    url(r'^scenario/(?P<scenario_id>\d*)/workflow/(?P<workflow_id>\d+)/loggings/step/(?P<step>\d+)$',
+        LoggingView.as_view(),
+        name='flooding_worker_workflow_logging'),
+    url(r'^scenario/(?P<scenario_id>\d*)/workflow/(?P<workflow_id>\d+)/task/(?P<task_id>\d*)/loggings/$',
+        LoggingView.as_view(),
+        name='flooding_worker_workflow_task_logging'),
+    url(r'^scenario/(?P<scenario_id>\d*)/workflow/(?P<workflow_id>\d+)/task/(?P<task_id>\d*)/loggings/stap/(?P<step>\d+)$',
+        LoggingView.as_view(),
+        name='flooding_worker_workflow_task_logging'),
     (r'^admin/', include(admin.site.urls)),
     )