+++ flooding_worker/worker/action.py
@@ -3,12 +3,21 @@
 
 import time
 import simplejson
-import logging
 
 from django.conf import settings
 
 
 class Action(object):
+
+    CREATED = u'CREATED'
+    QUEUED = u'QUEUED'
+    STARTED = u'STARTED'
+    SUCCESS = u'SUCCESS'
+    FAILED = u'FAILED'
+
+    ALIVE = u'ALIVE'
+    DOWN = u'DOWN'
+    BUSY = u'BUSY'
 
     def __init__(self):
         self.body = None
@@ -44,7 +53,7 @@ class Action(object):
         """Sets logging info into body."""
         self.body["message"] = str(message)
         self.body["curr_log_level"] = log_level
-        self.body["event_time"] = time.time()
+        self.body["time"] = time.time()
 
     def retrieve_queue_options(self, task_code):
         """Retrieves queue info from brokerconfig file."""
@@ -52,27 +61,22 @@ class Action(object):
 
     def root_queues(self):
         """Retrieve root queues from task's body."""
-        queues = []
         instruction = self.body["instruction"]
-        for (queue_code, parent_code) in instruction.iteritems():
-            if queue_code == parent_code:
-                queues.append(queue_code)
-        return queues
+        return [queue_code for queue_code, parent_code in instruction.iteritems()
+                if queue_code == parent_code]
 
     def next_queues(self):
         """
-        Recovers queues(s) of next task(s)
+        Recovers queue(s) of next task(s)
         by increasing the sequence.
         """
         instruction = self.body["instruction"]
         current_queue = self.body["curr_task_code"]
-        queues = []
-        for (queue_code, parent_code) in instruction.iteritems():
-            if queue_code == parent_code:
-                continue
-            if current_queue == parent_code:
-                queues.append(queue_code)
-        return queues
+        return [queue_code for queue_code, parent_code in instruction.iteritems()
+                if queue_code != parent_code and current_queue == parent_code]
 
     def set_current_task(self, queue):
         self.body["curr_task_code"] = queue
+
+    def set_task_status(self, status):
+        self.body["status"] = status